Apply code review suggestions

+++ Reemit.Decompiler.Clr/Methods/CorILMethodSectionFlags.cs
@@ -4,7 +4,7 @@ namespace Reemit.Decompiler.Clr.Methods;
 /// Based on method data section flags defined in II.25.4.5 Method data section
 /// </summary>
 [Flags]
-public enum CorILMethodSectionFlags : uint
+public enum CorILMethodSectionFlags : byte
 {
     EHTable = 0x1,
     OptILTable = 0x2,
